Move header to types file
