+++ src/components/content-header-card/types.ts
@@ -1,0 +1,38 @@
+import { Badge } from 'components/badge-list/types'
+import { ProductSlug } from 'types/products'
+
+export interface DropdownItem {
+	text: string
+	href: string
+}
+
+export interface Dropdown {
+	text: string
+	items: Array<DropdownItem>
+}
+
+export interface Link {
+	text: string
+	href: string
+	icon?: React.ReactElement
+}
+
+export interface Button {
+	text: string
+	icon?: JSX.IntrinsicElements['svg']
+	isPrimary?: boolean
+	onClick: () => void
+}
+
+export interface ContentHeaderCardProps {
+	className?: string
+	icon?: Exclude<ProductSlug, 'sentinel'>
+	title: string
+	attribution?: string
+	description?: string
+	note?: string
+	badges?: Array<Badge>
+	dropdown?: Dropdown
+	links?: Array<Link>
+	buttons?: Array<Button>
+}
